refactor visualisation utils
